Add most of lab 1
Based on the workflow for submission, I'm dumping all of the existing
files since the module lecturer will likely be checking out each
individual branch and doesn't need to see the same README and lab sheets
over and over and over.

This doesn't include the third part of the lab, and as it seems we'll be
re-using it in the future I am considering re-writing the calculator to
use css grids and or flexbox to be cleaner.