Add new branch for lab 2
I'm not finished with lab 1 yet, but I'm setting up the scaffolding for
the second lab with this branch/commit. Hopefully I can get the
automata exercise done and pushed tonight so that I can catch up with
this lab work - I've decided to go forward with the version of the
calculator which uses CSS grid.